raster-nodes: move `CubicSplines` to separate mod
